notification_service/worker/requirements.txt
kafka-python == 2.3.0
aiokafka==0.13.0
pydantic-settings==2.11.0
pydantic==2.12.0
aiohttp==3.12.15
aiosmtplib==5.1.0
motor==3.7.1